requirements.txt
streamlit
python-dotenv
google-generativeai
youtube-transcript-api